# Speed up android builds on Travis