--- a/.travis.yml
+++ b/.travis.yml
@@ -92,7 +92,7 @@ script:
   - if [[ "$TEST_TYPE" == "ios" ]]; then npm run test:ios; fi
 
   # android-only tests - building and specs
-  - if [[ "$TEST_TYPE" = android ]]; then npm run build:android:macos; fi
+  - if [[ "$TEST_TYPE" == "android" ]]; then npm run build:android:unix; fi
 
 # ping slack with status
 notifications: