--- a/android/app/build.gradle
+++ b/android/app/build.gradle
@@ -126,6 +126,12 @@ android {
             }
         }
     }
+    android {
+        dexOptions {
+            // Skip pre-dexing when running on Travis CI or when disabled via -Dpre-dex=false.
+            preDexLibraries = preDexEnabled && !travisBuild
+        }
+    }
 }
 
 dependencies {