--- a/android/build.gradle
+++ b/android/build.gradle
@@ -12,6 +12,12 @@ buildscript {
     }
 }
 
+ext {
+    travisBuild = System.getenv("TRAVIS") == "true"
+    // allows for -Dpre-dex=false to be set
+    preDexEnabled = "true".equals(System.getProperty("pre-dex", "true"))
+}
+
 allprojects {
     repositories {
         mavenLocal()