I enabled a more … selective build matrix on Travis.

- JS builds only need JS tooling, so we use Travis' docker-based containers for fast boot. this trims the runtime for js-tests from around 3-4 minutes to about 45-50 seconds.

- Android builds have a dedicated `android` environment that we can take advantage of. use it. this drops the time from ~20 minutes to ~9 minutes.

- I added the tooling to support running an android emulator on the new builds. however, it takes about 20 minutes to run the android tests on this system because of the emulator. when we decide to build more tests for the emulator, this will be useful.

- OS X / iOS builds still use the same image.

None of this affected Appveyor, whose builds are currently averaging 7 minutes (windows, android-only).

Oh, and the linux image-based builds also start up faster 😉. They usually run about a minute ahead of the OS X-based images.

---

### Before ([build 138][])

- **JS:** 1 min, 59 sec
- **iOS:** 8 min, 56 sec
- **Android:** 20 min, 25 sec
- **Total:** 31 min, 20 sec

### After ([build 163][])

- **JS:** 0 min, 48 sec
- **iOS:** 10 min, 44 sec
- **Android:** 9 min, 5 sec
- **Total:** 20 min, 37 sec

[build 138]: https://travis-ci.org/StoDevX/AAO-React-Native/builds/152887135
[build 163]: https://travis-ci.org/StoDevX/AAO-React-Native/builds/152953913